% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-visualization.R
\name{add_boundary_overlay}
\alias{add_boundary_overlay}
\title{Add boundary overlay to existing plot}
\usage{
add_boundary_overlay(region_boundary, verbose)
}
\description{
Add boundary overlay to existing plot
}
\keyword{internal}
