% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{add_crop_statistics_overlay}
\alias{add_crop_statistics_overlay}
\title{Add crop statistics overlay}
\usage{
add_crop_statistics_overlay(plot_obj, cdl_raster, crop_codes, crop_names)
}
\description{
Add crop statistics overlay
}
\keyword{internal}
