% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{add_lines_to_leaflet}
\alias{add_lines_to_leaflet}
\title{Add lines to leaflet -}
\usage{
add_lines_to_leaflet(
  map_obj,
  spatial_data,
  fill_variable,
  popup_vars,
  color_scheme
)
}
\description{
Add lines to leaflet -
}
\keyword{internal}
