% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{add_points_to_leaflet}
\alias{add_points_to_leaflet}
\title{Add points to leaflet -}
\usage{
add_points_to_leaflet(
  map_obj,
  spatial_data,
  fill_variable,
  popup_vars,
  color_scheme,
  cluster_points
)
}
\description{
Add points to leaflet -
}
\keyword{internal}
