% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-temporal-analysis.R
\name{analyze_seasonal_patterns}
\alias{analyze_seasonal_patterns}
\title{Analyze seasonal patterns}
\usage{
analyze_seasonal_patterns(rasters, dates)
}
\arguments{
\item{rasters}{List of rasters}

\item{dates}{Vector of dates}
}
\value{
List with seasonal analysis results
}
\description{
Internal function to analyze seasonal patterns in temporal data.
}
\keyword{internal}
