% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-visualization.R
\name{apply_color_scheme_ggplot}
\alias{apply_color_scheme_ggplot}
\title{Apply color scheme to ggplot}
\usage{
apply_color_scheme_ggplot(plot, color_scheme, aesthetic = "fill")
}
\description{
Apply color scheme to ggplot
}
\keyword{internal}
