% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-vegetation-indices.R
\name{apply_quality_filter}
\alias{apply_quality_filter}
\title{Apply quality filter to remove outliers}
\usage{
apply_quality_filter(ndvi, verbose)
}
\description{
Apply quality filter to remove outliers
}
\keyword{internal}
