% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-vegetation-indices.R
\name{apply_temporal_smoothing}
\alias{apply_temporal_smoothing}
\title{Apply temporal smoothing}
\usage{
apply_temporal_smoothing(ndvi_stack, verbose)
}
\description{
Apply temporal smoothing
}
\keyword{internal}
