% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{auto_detect_method}
\alias{auto_detect_method}
\title{Auto-detect best spatial join method}
\usage{
auto_detect_method(source_type, target_type, verbose = FALSE)
}
\arguments{
\item{source_type}{Type of source data}

\item{target_type}{Type of target data}

\item{verbose}{Print messages}
}
\value{
Character string of recommended method
}
\description{
Auto-detect best spatial join method
}
\keyword{internal}
