% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-vegetation-indices.R
\name{auto_detect_spectral_bands}
\alias{auto_detect_spectral_bands}
\title{Auto-detect spectral bands}
\usage{
auto_detect_spectral_bands(spectral_data, layer_names, verbose)
}
\description{
Auto-detect spectral bands
}
\keyword{internal}
