% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-visualization.R
\name{auto_detect_title}
\alias{auto_detect_title}
\title{Auto-detect title for quick mapping}
\usage{
auto_detect_title(spatial_data, variable)
}
\description{
Auto-detect title for quick mapping
}
\keyword{internal}
