% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-terrain-analysis.R
\name{calculate_contributing_area}
\alias{calculate_contributing_area}
\title{Calculate contributing area}
\usage{
calculate_contributing_area(dem)
}
\arguments{
\item{dem}{Digital elevation model}
}
\value{
SpatRaster of contributing area
}
\description{
Internal function to calculate contributing area for wetness index.
}
\keyword{internal}
