% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-vegetation-indices.R
\name{calculate_index_by_type}
\alias{calculate_index_by_type}
\title{Calculate index by type}
\usage{
calculate_index_by_type(
  index_type,
  red,
  nir,
  blue,
  green,
  swir1,
  swir2,
  red_edge,
  coastal,
  nir2,
  verbose
)
}
\description{
Calculate index by type
}
\keyword{internal}
