% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-interpolation.R
\name{calculate_skewness}
\alias{calculate_skewness}
\title{Calculate skewness for data assessment}
\usage{
calculate_skewness(x)
}
\description{
Calculate skewness for data assessment
}
\keyword{internal}
