% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-temporal-analysis.R
\name{calculate_temporal_statistics}
\alias{calculate_temporal_statistics}
\title{Calculate temporal statistics}
\usage{
calculate_temporal_statistics(rasters, dates)
}
\arguments{
\item{rasters}{List of rasters}

\item{dates}{Vector of dates}
}
\value{
List with temporal statistics
}
\description{
Internal function to calculate temporal statistics.
}
\keyword{internal}
