% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-temporal-analysis.R
\name{calculate_temporal_trend}
\alias{calculate_temporal_trend}
\title{Calculate temporal trend using linear regression}
\usage{
calculate_temporal_trend(rasters, dates)
}
\arguments{
\item{rasters}{List of rasters}

\item{dates}{Vector of dates}
}
\value{
List with trend analysis results
}
\description{
Internal function to calculate pixel-wise temporal trends.
}
\keyword{internal}
