% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-vegetation-indices.R
\name{calculate_vegetation_statistics}
\alias{calculate_vegetation_statistics}
\title{Calculate vegetation statistics}
\usage{
calculate_vegetation_statistics(vegetation_indices, indices, verbose)
}
\description{
Calculate vegetation statistics
}
\keyword{internal}
