% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{classify_data_input}
\alias{classify_data_input}
\title{Classify data input type}
\usage{
classify_data_input(data_input, verbose = FALSE)
}
\arguments{
\item{data_input}{Input data of various types}

\item{verbose}{Print diagnostic messages}
}
\value{
List with type, class, and data object
}
\description{
Classify data input type
}
\keyword{internal}
