% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{classify_spatial_data}
\alias{classify_spatial_data}
\title{Classify spatial data type and load data}
\usage{
classify_spatial_data(data_input, verbose = FALSE)
}
\arguments{
\item{data_input}{Input data (file path, directory, or spatial object)}

\item{verbose}{Print progress messages}
}
\value{
List with data, type, class, and path information
}
\description{
Classify spatial data type and load data
}
\keyword{internal}
