% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-interpolation.R
\name{compile_interpolation_results}
\alias{compile_interpolation_results}
\title{Compile interpolation results into requested format}
\usage{
compile_interpolation_results(
  interpolation_results,
  original_data,
  target_variables,
  output_format,
  verbose
)
}
\description{
Compile interpolation results into requested format
}
\keyword{internal}
