% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-visualization.R
\name{create_base_plot_map}
\alias{create_base_plot_map}
\title{Create base plot map as fallback}
\usage{
create_base_plot_map(spatial_sf, fill_variable, map_type, title, verbose)
}
\description{
Create base plot map as fallback
}
\keyword{internal}
