% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-correlation.R
\name{create_correlation_plots}
\alias{create_correlation_plots}
\title{Create correlation plots}
\usage{
create_correlation_plots(correlation_matrix, output_folder)
}
\arguments{
\item{correlation_matrix}{Correlation matrix}

\item{output_folder}{Output directory}
}
\description{
Internal function to create correlation plots and heatmaps.
}
\keyword{internal}
