% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{create_crop_area_map}
\alias{create_crop_area_map}
\title{Create crop area map -}
\usage{
create_crop_area_map(cdl_raster, crop_codes, crop_names, boundary)
}
\description{
Create crop area map -
}
\keyword{internal}
