% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{create_crop_diversity_map}
\alias{create_crop_diversity_map}
\title{Create crop diversity map -}
\usage{
create_crop_diversity_map(cdl_raster, crop_codes, boundary)
}
\description{
Create crop diversity map -
}
\keyword{internal}
