% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{create_dominant_crop_map}
\alias{create_dominant_crop_map}
\title{Create dominant crop map -}
\usage{
create_dominant_crop_map(cdl_raster, crop_codes, crop_names, boundary)
}
\description{
Create dominant crop map -
}
\keyword{internal}
