% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-visualization.R
\name{create_ggplot_map_safe}
\alias{create_ggplot_map_safe}
\title{Create ggplot map safely}
\usage{
create_ggplot_map_safe(
  spatial_sf,
  fill_variable,
  map_type,
  region_boundary,
  color_scheme,
  title,
  point_size,
  verbose
)
}
\description{
Create ggplot map safely
}
\keyword{internal}
