% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-visualization.R
\name{create_interactive_map_safe}
\alias{create_interactive_map_safe}
\title{Create interactive map safely}
\usage{
create_interactive_map_safe(
  spatial_sf,
  fill_variable,
  map_type,
  title,
  verbose
)
}
\description{
Create interactive map safely
}
\keyword{internal}
