% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{create_overlay_comparison}
\alias{create_overlay_comparison}
\title{Create overlay comparison}
\usage{
create_overlay_comparison(data1, data2, titles, color_scheme)
}
\description{
Create overlay comparison
}
\keyword{internal}
