% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-visualization.R
\name{create_raster_map_reliable}
\alias{create_raster_map_reliable}
\title{Create reliable raster map}
\usage{
create_raster_map_reliable(
  raster_data,
  region_boundary,
  color_scheme,
  title,
  output_file,
  verbose
)
}
\description{
Create reliable raster map
}
\keyword{internal}
