% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-water-quality.R
\name{detect_coordinate_columns}
\alias{detect_coordinate_columns}
\title{Detect coordinate columns automatically}
\usage{
detect_coordinate_columns(water_df, verbose)
}
\description{
Detect coordinate columns automatically
}
\keyword{internal}
