% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-visualization.R
\name{detect_geometry_type}
\alias{detect_geometry_type}
\title{Detect geometry type automatically}
\usage{
detect_geometry_type(spatial_sf)
}
\description{
Detect geometry type automatically
}
\keyword{internal}
