% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{detect_huc_level}
\alias{detect_huc_level}
\title{Detect HUC level from column name or data}
\usage{
detect_huc_level(column_name, data_values)
}
\arguments{
\item{column_name}{Name of the HUC column}

\item{data_values}{Sample HUC values}
}
\value{
Character string of HUC type (e.g., "huc08", "huc12")
}
\description{
Detect HUC level from column name or data
}
\keyword{internal}
