% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-temporal-analysis.R
\name{detect_temporal_changes}
\alias{detect_temporal_changes}
\title{Detect temporal changes between periods}
\usage{
detect_temporal_changes(rasters, dates)
}
\arguments{
\item{rasters}{List of rasters}

\item{dates}{Vector of dates}
}
\value{
List with change detection results
}
\description{
Internal function to detect changes between time periods.
}
\keyword{internal}
