% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-water-quality.R
\name{detect_water_quality_columns}
\alias{detect_water_quality_columns}
\title{Detect water quality data columns intelligently}
\usage{
detect_water_quality_columns(
  water_sf,
  variable,
  date_column,
  station_id_col,
  verbose
)
}
\description{
Detect water quality data columns intelligently
}
\keyword{internal}
