% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-interpolation.R
\name{execute_nn_interpolation}
\alias{execute_nn_interpolation}
\title{Execute nearest neighbor interpolation}
\usage{
execute_nn_interpolation(complete_data, variable, target_grid, verbose)
}
\description{
Execute nearest neighbor interpolation
}
\keyword{internal}
