% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-vegetation-indices.R
\name{extract_bands_from_raster}
\alias{extract_bands_from_raster}
\title{Extract bands from multi-band raster}
\usage{
extract_bands_from_raster(spectral_data, auto_detect, band_names, verbose)
}
\description{
Extract bands from multi-band raster
}
\keyword{internal}
