% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-workflows.R
\name{generate_html_summary}
\alias{generate_html_summary}
\title{Generate HTML summary report}
\usage{
generate_html_summary(summary_report, results, output_file)
}
\description{
Generate HTML summary report
}
\keyword{internal}
