% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{geocode_entities}
\alias{geocode_entities}
\title{Geocode geographic entities to coordinates}
\usage{
geocode_entities(data_input, entity_info, verbose = FALSE)
}
\arguments{
\item{data_input}{Data frame with geographic entity column}

\item{entity_info}{List from detect_geographic_entities}

\item{verbose}{Print messages}
}
\value{
sf object with geocoded locations
}
\description{
Geocode geographic entities to coordinates
}
\keyword{internal}
