% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-vegetation-indices.R
\name{get_index_required_bands}
\alias{get_index_required_bands}
\title{Get required bands for each vegetation index}
\usage{
get_index_required_bands(index_type)
}
\description{
Get required bands for each vegetation index
}
\keyword{internal}
