% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-water-indices.R
\name{get_interpretation_guidelines}
\alias{get_interpretation_guidelines}
\title{Get interpretation guidelines}
\usage{
get_interpretation_guidelines(indices)
}
\description{
Get interpretation guidelines
}
\keyword{internal}
