% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-visualization.R
\name{get_reliable_colors}
\alias{get_reliable_colors}
\title{Get reliable colors for terra plotting}
\usage{
get_reliable_colors(color_scheme, n = 100)
}
\description{
Get reliable colors for terra plotting
}
\keyword{internal}
