% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-water-indices.R
\name{get_satellite_band_info}
\alias{get_satellite_band_info}
\title{Get satellite band information}
\usage{
get_satellite_band_info(indices)
}
\description{
Get satellite band information
}
\keyword{internal}
