% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{get_summary_function}
\alias{get_summary_function}
\title{Get summary function for terra operations}
\usage{
get_summary_function(summary_function)
}
\arguments{
\item{summary_function}{Character name of function}
}
\value{
Function object
}
\description{
Get summary function for terra operations
}
\keyword{internal}
