% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-visualization.R
\name{get_terra_colors}
\alias{get_terra_colors}
\title{Get terra colors for plotting}
\usage{
get_terra_colors(scheme = "viridis", n = 100)
}
\description{
Get terra colors for plotting
}
\keyword{internal}
