% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-water-indices.R
\name{get_water_index_requirements}
\alias{get_water_index_requirements}
\title{Get water index requirements}
\usage{
get_water_index_requirements()
}
\description{
Get water index requirements
}
\keyword{internal}
