% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-water-indices.R
\name{handle_water_index_edge_cases}
\alias{handle_water_index_edge_cases}
\title{Handle edge cases for water index calculations}
\usage{
handle_water_index_edge_cases(index, index_type, verbose)
}
\description{
Handle edge cases for water index calculations
}
\keyword{internal}
