% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{integrate_multiple_datasets}
\alias{integrate_multiple_datasets}
\title{Integrate multiple datasets}
\usage{
integrate_multiple_datasets(
  vector_data,
  raster_datasets,
  region_boundary = NULL,
  output_format = "sf"
)
}
\description{
Integrate multiple datasets
}
\keyword{internal}
