% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-interpolation.R
\name{load_and_process_spatial_data}
\alias{load_and_process_spatial_data}
\title{Load and process spatial data for interpolation}
\usage{
load_and_process_spatial_data(spatial_data, coord_cols, verbose)
}
\description{
Load and process spatial data for interpolation
}
\keyword{internal}
