% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-vegetation-indices.R
\name{load_and_stack_bands}
\alias{load_and_stack_bands}
\title{Load and stack individual band files}
\usage{
load_and_stack_bands(file_list, band_names = NULL, verbose = FALSE)
}
\description{
Load and stack individual band files
}
\keyword{internal}
