% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-vegetation-indices.R
\name{load_bands_from_directory}
\alias{load_bands_from_directory}
\title{Load bands from directory}
\usage{
load_bands_from_directory(directory, band_names = NULL, verbose = FALSE)
}
\description{
Load bands from directory
}
\keyword{internal}
