% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-water-quality.R
\name{load_single_water_dataset}
\alias{load_single_water_dataset}
\title{Load single water quality dataset}
\usage{
load_single_water_dataset(data_input, coord_cols, verbose)
}
\description{
Load single water quality dataset
}
\keyword{internal}
