% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{map_custom_band_names}
\alias{map_custom_band_names}
\title{Map custom band names}
\usage{
map_custom_band_names(spectral_data, band_names, verbose = FALSE)
}
\description{
Map custom band names
}
\keyword{internal}
