% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-vegetation-indices.R
\name{mask_invalid_values}
\alias{mask_invalid_values}
\title{Mask invalid values based on index type}
\usage{
mask_invalid_values(index, index_type, verbose)
}
\description{
Mask invalid values based on index type
}
\keyword{internal}
