% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-water-indices.R
\name{mask_invalid_water_values}
\alias{mask_invalid_water_values}
\title{Mask invalid values for water indices}
\usage{
mask_invalid_water_values(index, index_type, verbose)
}
\description{
Mask invalid values for water indices
}
\keyword{internal}
