% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{match_rasters_by_date}
\alias{match_rasters_by_date}
\title{Match rasters by date}
\usage{
match_rasters_by_date(
  red_data,
  nir_data,
  date_patterns = NULL,
  verbose = FALSE
)
}
\description{
Match rasters by date
}
\keyword{internal}
