% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-interpolation.R
\name{perform_cross_validation}
\alias{perform_cross_validation}
\title{Perform cross-validation for interpolation accuracy}
\usage{
perform_cross_validation(
  spatial_sf,
  target_variables,
  method,
  cv_folds,
  power,
  max_distance,
  min_points,
  max_points,
  verbose
)
}
\description{
Perform cross-validation for interpolation accuracy
}
\keyword{internal}
