% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{perform_extract_join}
\alias{perform_extract_join}
\title{Perform extract join (vector to raster)}
\usage{
perform_extract_join(
  source_info,
  target_info,
  summary_function,
  buffer_distance,
  crs_target,
  na_strategy,
  verbose
)
}
\description{
Perform extract join (vector to raster)
}
\keyword{internal}
