% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{perform_resample_join}
\alias{perform_resample_join}
\title{Perform resample join (raster to raster)}
\usage{
perform_resample_join(
  source_info,
  target_info,
  scale_factor,
  summary_function,
  crs_target,
  verbose
)
}
\description{
Perform resample join (raster to raster)
}
\keyword{internal}
