% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-visualization.R, R/15-helpers.R
\name{process_vector_data}
\alias{process_vector_data}
\title{Process vector data from data frame}
\usage{
process_vector_data(data_input, coord_cols = c("lon", "lat"))

process_vector_data(data_input, coord_cols = c("lon", "lat"))
}
\arguments{
\item{data_input}{Data frame with potential coordinate columns}

\item{coord_cols}{Coordinate column names}
}
\value{
sf object
}
\description{
Process vector data from data frame

Process vector data from data frame
}
\keyword{internal}
