% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-visualization.R
\name{quick_map}
\alias{quick_map}
\title{Quick map function - one-line mapping with auto-detection}
\usage{
quick_map(spatial_data, variable = NULL, title = NULL, ...)
}
\arguments{
\item{spatial_data}{Any spatial data}

\item{variable}{Variable to visualize (optional, auto-detected)}

\item{title}{Map title (optional)}

\item{...}{Additional arguments passed to create_spatial_map}
}
\value{
Map object
}
\description{
Ultra-simple function for quick spatial mapping. Auto-detects data type and creates appropriate map.
}
\examples{
\dontrun{
# These examples require external data files not included with the package
quick_map("data.shp")
quick_map(my_raster)
quick_map(points_data, interactive = TRUE)
}

}
