% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-workflows.R, R/15-helpers.R
\name{run_enhanced_mosaic_workflow}
\alias{run_enhanced_mosaic_workflow}
\title{Run enhanced mosaic workflow}
\usage{
run_enhanced_mosaic_workflow(config, output_folder)

run_enhanced_mosaic_workflow(config, output_folder)
}
\description{
Run enhanced mosaic workflow

Run enhanced mosaic workflow - IMPLEMENTED
}
\keyword{internal}
