% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{run_multi_dataset_workflow}
\alias{run_multi_dataset_workflow}
\title{Run multi-dataset workflow - IMPLEMENTED}
\usage{
run_multi_dataset_workflow(config, output_folder)
}
\description{
Run multi-dataset workflow - IMPLEMENTED
}
\keyword{internal}
