% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-visualization.R
\name{save_interactive_map}
\alias{save_interactive_map}
\title{Save interactive map to file}
\usage{
save_interactive_map(map_object, output_file, verbose)
}
\description{
Save interactive map to file
}
\keyword{internal}
