% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-interpolation.R
\name{save_interpolation_results}
\alias{save_interpolation_results}
\title{Save interpolation results to file}
\usage{
save_interpolation_results(results, output_file, output_format, verbose)
}
\description{
Save interpolation results to file
}
\keyword{internal}
