% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-visualization.R
\name{save_static_map}
\alias{save_static_map}
\title{Save static map with ggplot2}
\usage{
save_static_map(plot_object, output_file, verbose)
}
\description{
Save static map with ggplot2
}
\keyword{internal}
