% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-temporal-analysis.R
\name{save_temporal_results}
\alias{save_temporal_results}
\title{Save temporal analysis results}
\usage{
save_temporal_results(result, output_folder, analysis_type)
}
\arguments{
\item{result}{Temporal analysis results}

\item{output_folder}{Output directory}

\item{analysis_type}{Type of analysis}
}
\description{
Internal function to save temporal analysis results to files.
}
\keyword{internal}
