% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{save_vegetation_analysis_results}
\alias{save_vegetation_analysis_results}
\title{Save vegetation analysis results}
\usage{
save_vegetation_analysis_results(
  analysis_results,
  vegetation_indices,
  output_folder,
  crop_type,
  verbose = FALSE
)
}
\description{
Save vegetation analysis results
}
\keyword{internal}
