% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-water-quality.R
\name{save_water_quality_results}
\alias{save_water_quality_results}
\title{Save water quality analysis results}
\usage{
save_water_quality_results(
  water_sf,
  variable,
  statistics,
  threshold_results,
  spatial_analysis,
  temporal_analysis,
  region_boundary,
  river_network,
  output_folder,
  verbose
)
}
\description{
Save water quality analysis results
}
\keyword{internal}
