% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-testing.R
\name{test_function_availability}
\alias{test_function_availability}
\title{Test individual function existence}
\usage{
test_function_availability(verbose = FALSE)
}
\arguments{
\item{verbose}{Print detailed messages}
}
\value{
List of function availability
}
\description{
Helper function to test if core functions exist and are callable.
}
