% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-testing.R
\name{test_package_minimal}
\alias{test_package_minimal}
\title{Test package with minimal complexity}
\usage{
test_package_minimal(verbose = FALSE)
}
\arguments{
\item{verbose}{Print messages}
}
\value{
Logical indicating basic functionality
}
\description{
Ultra-minimal test that only checks the most basic functionality.
Designed to always pass if the package is minimally functional.
}
