% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-validation-helpers.R
\name{validate_method}
\alias{validate_method}
\title{Validate method parameter}
\usage{
validate_method(method, valid_methods, param_name = "method")
}
\arguments{
\item{method}{Method to validate}

\item{valid_methods}{Vector of valid methods}

\item{param_name}{Parameter name for error messages}
}
\value{
TRUE if valid, stops with error if invalid
}
\description{
Validate method parameter
}
\keyword{internal}
