% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{validate_method_compatibility}
\alias{validate_method_compatibility}
\title{Validate method compatibility}
\usage{
validate_method_compatibility(method, source_type, target_type)
}
\arguments{
\item{method}{Selected method}

\item{source_type}{Source data type}

\item{target_type}{Target data type}
}
\description{
Validate method compatibility
}
\keyword{internal}
