% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-validation-helpers.R
\name{validate_numeric_range}
\alias{validate_numeric_range}
\title{Validate numeric range}
\usage{
validate_numeric_range(
  x,
  min_val = NULL,
  max_val = NULL,
  param_name = "parameter",
  allow_null = FALSE
)
}
\arguments{
\item{x}{Numeric value to validate}

\item{min_val}{Minimum allowed value}

\item{max_val}{Maximum allowed value}

\item{param_name}{Parameter name for error messages}

\item{allow_null}{Allow NULL values}
}
\value{
TRUE if valid, stops with error if invalid
}
\description{
Validate numeric range
}
\keyword{internal}
