% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-validation-helpers.R
\name{validate_raster_input}
\alias{validate_raster_input}
\title{Validate raster input}
\usage{
validate_raster_input(x, param_name = "raster_data", allow_null = FALSE)
}
\arguments{
\item{x}{Input to validate}

\item{param_name}{Parameter name for error messages}

\item{allow_null}{Allow NULL values}
}
\value{
TRUE if valid, stops with error if invalid
}
\description{
Validate raster input
}
\keyword{internal}
