% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-vegetation-indices.R
\name{validate_required_bands}
\alias{validate_required_bands}
\title{Validate required bands for specific indices}
\usage{
validate_required_bands(
  index_type,
  blue,
  green,
  swir1,
  swir2,
  red_edge,
  coastal,
  nir2,
  verbose = FALSE
)
}
\description{
Validate required bands for specific indices
}
\keyword{internal}
