% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-validation-helpers.R
\name{validate_vector_input}
\alias{validate_vector_input}
\title{Validate vector input}
\usage{
validate_vector_input(x, param_name = "vector_data", allow_null = FALSE)
}
\arguments{
\item{x}{Input to validate}

\item{param_name}{Parameter name for error messages}

\item{allow_null}{Allow NULL values}
}
\value{
TRUE if valid, stops with error if invalid
}
\description{
Validate vector input
}
\keyword{internal}
