% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-helpers.R
\name{validate_vegetation_analysis}
\alias{validate_vegetation_analysis}
\title{Validate vegetation analysis}
\usage{
validate_vegetation_analysis(
  vegetation_indices,
  reference_data,
  verbose = FALSE
)
}
\description{
Validate vegetation analysis
}
\keyword{internal}
