% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-water-indices.R
\name{validate_water_index_output}
\alias{validate_water_index_output}
\title{Validate water index output}
\usage{
validate_water_index_output(index, index_type, verbose)
}
\description{
Validate water index output
}
\keyword{internal}
